added helpers in Header, fix bad practice async in useEffect, change function for const and fix missing key in AnimeSections
