+++ src/components/Header.tsx
@@ -1,4 +1,4 @@
-import { Text, Link, IconButton, Input, Image, InputGroup, InputLeftElement, Modal, ModalContent, ModalOverlay, useDisclosure, Box, Button, Flex, Heading, Icon, Avatar, useToast, ListItem, AspectRatio, Divider, CircularProgress } from "@chakra-ui/react";
+import { Text, Link, IconButton, Input, Image, InputGroup, InputLeftElement, Modal, ModalContent, ModalOverlay, Box, Button, Flex, Heading, Icon, useToast, Divider, CircularProgress } from "@chakra-ui/react";
 import { SearchIcon } from "@chakra-ui/icons";
 import { ImGithub, ImLinkedin, ImInstagram } from 'react-icons/im'
 import { FaTools } from "react-icons/fa";
@@ -6,60 +6,119 @@ import { AiOutlineSearch } from "react-icons/ai";
 import { GiHamburgerMenu } from "react-icons/gi";
 import DarkModeToggle from './ThemeSwitch';
 import { Link as RLink } from 'react-router-dom'
-import { SetStateAction, useEffect, useRef, useState } from 'react';
-import { searchAnime } from "../services/ApiServices";
-import { Anime } from "../types/interface";
+import { useEffect } from 'react';
+import helpers from './helpers/Header/helpers';
 import '../App.css';
 
-
-function MyHeader() {
-    const { isOpen, onOpen, onClose } = useDisclosure();
+const MyHeader = () => {
     const toast = useToast();
-    const [height, setHeight] = useState('0px');
-    const [isClicked, setClicked] = useState<boolean>(false);
-    const [searchResult, setsearchResult] = useState<Anime[]>()
-    const [query, setQuery] = useState<string>('');
-    const [loading, setLoading] = useState<boolean>(false);
+    //All the functions I put in helpers, this going to make a little easier to read the code, bc I can just call the function instead of writing the whole function in here
+    const {
+        search,
+        handleHamburger,
+        handleClose,
+        handleOpen,
+        handleClickSearch,
+        setsearchResult,
+        setQuery,
+        handleSearch,
+        onOpen,
+        onClose,
+        loading,
+        searchResult,
+        isOpen,
+        query,
+        isClicked,
+        height,
+    } = helpers();
 
+
+    // Dont use async function in useEffect its a bad practice, so I use a function inside the useEffect and call the async function inside the function
     useEffect(() => {
-        console.log('EF')
-        const search = async () => {
-            setLoading(true);
-            const res = await searchAnime(query);
-            console.log('res', res);
-            setsearchResult(res)
-            setLoading(false);
-        }
         search();
-
     }, [query])
 
-    function handleHamburger() {
 
-        setClicked(isClicked ? false : true);
-        setHeight(height == '0px' ? '300px' : '0px')
+    // I move this up bc I thinks its more easy to read and change function for arrow function, plus I consider to move all of this functions in another folder, I think this overhead the code
+
+    const SearchBar = () => {
+        return (
+            <InputGroup onClick={onOpen} size="lg" color="gray.800">
+                <InputLeftElement pointerEvents="none" children={<SearchIcon color="gray.400" />} />
+                <Input type="text" placeholder="Search Anime..." />
+            </InputGroup>
+        );
     }
 
-    const handleClose = () => {
-        setsearchResult([]); // clear the query state variable
-        onClose(); // call the original onClose function
-    };
-
-    const handleOpen = () =>{
-        setsearchResult([]); // clear the query state variable
-        onOpen(); // 
-    }
-    const handleClickSearch = ()=>{
-        setsearchResult([]); 
+    const SearchBarModal = () => {
+        return (
+            <InputGroup onKeyDown={handleSearch} onClick={onOpen} size="lg" color="gray.800">
+                <InputLeftElement pointerEvents="none" children={<SearchIcon color="gray.400" />} />
+                <Input _dark={{ textColor: 'white' }} type="text" placeholder="Search Anime..." />
+            </InputGroup>
+        );
     }
 
-    const handleSearch = (event: React.KeyboardEvent<HTMLInputElement>) => {
-        if (event.key === "Enter") {
-            setsearchResult([]);
-            const inputElement = event.target as HTMLInputElement;
-            console.log(inputElement.value)
-            setQuery(inputElement.value);
-        }
+    const  MyModal = () => {
+        return (
+            <Modal isOpen={isOpen} onClose={handleClose}>
+                <ModalOverlay
+                    bg="blackAlpha.300"
+                    backdropInvert="80%"
+                    backdropFilter="blur(2px) hue-rotate(90deg)"
+                />
+                <ModalContent bg="rgba(255, 255, 255, 0.8)" /* semi-transparent background color */
+                    backdropFilter="blur(10px)" /* for Chrome */
+                    _dark={{ bg: 'rgba(0, 0, 0, 0.8)' }} maxW={{ base: '300px', md: '540px' }} >
+                    <SearchBarModal />
+                    <Flex maxH="80vh" flexDir={'column'} gap={3} overflowY={'scroll'}>
+                        {loading ?
+                            <Flex justifyContent={'center'}>
+                                <CircularProgress color={'blackAlpha.800'} />
+                            </Flex> :
+                            searchResult?.map((result) => (
+                                <>
+                                    <RLink onClick={onClose} to={`/anime/${result.id}`}>
+                                        <Flex gap={3} px={4} py={2}>
+                                            <Image src={result.image} alt={result.title?.romaji} height={100} width={'16'} fit={'cover'} />
+                                            <Flex flexDir={'column'}>
+                                                <Text fontSize={'md'} fontWeight={'semibold'} noOfLines={1}>{result.title?.romaji}</Text>
+                                                <Text fontSize={'sm'} fontWeight={'medium'} noOfLines={1}>{result.releaseDate}</Text>
+                                            </Flex>
+
+                                        </Flex>
+                                        <Divider></Divider>
+                                    </RLink>
+                                </>
+                            ))}
+                    </Flex>
+
+                </ModalContent>
+            </Modal>
+        );
+    }
+
+    function ConstructionBox() {
+        return (
+            <Box p={{ base: 8, md: 8 }} borderRadius="md">
+                <Flex flexDirection={'column'} gap={6}>
+                    <Box>
+                        <Flex alignItems="center" mb={4}>
+                            <Icon as={FaTools} fontSize="2xl" mr={2} />
+                            <Heading size={{ base: 'md', md: 'md' }}>Seacrh Feature Under Construction</Heading>
+                        </Flex>
+                        <Text fontSize="lg">
+                            We are currently working hard to bring you this search feature. Stay tuned for
+                            updates!
+                        </Text>
+                    </Box>
+                    <Button onClick={onClose} >
+                        Got It
+                    </Button>
+                </Flex>
+
+            </Box>
+        );
     };
 
 
@@ -162,96 +221,9 @@ function MyHeader() {
                     <Button w={'full'} bgColor={'transparent'}>Top Airing</Button>
                     <Button w={'full'} bgColor={'transparent'}>Recent Relase</Button>
                 </Flex>
-
-
             </Flex>
         </>
     );
-
-    function SearchBar() {
-        return (
-            <InputGroup onClick={onOpen} size="lg" color="gray.800">
-                <InputLeftElement pointerEvents="none" children={<SearchIcon color="gray.400" />} />
-                <Input type="text" placeholder="Search Anime..." />
-            </InputGroup>
-        );
-    }
-
-    function SearchBarModal() {
-        return (
-            <InputGroup onKeyDown={handleSearch} onClick={onOpen} size="lg" color="gray.800">
-                <InputLeftElement pointerEvents="none" children={<SearchIcon color="gray.400" />} />
-                <Input _dark={{textColor : 'white'}} type="text" placeholder="Search Anime..." />
-            </InputGroup>
-        );
-    }
-
-    function MyModal() {
-        return (
-            <Modal isOpen={isOpen} onClose={handleClose}>
-                <ModalOverlay
-                    bg="blackAlpha.300"
-                    backdropInvert="80%"
-                    backdropFilter="blur(2px) hue-rotate(90deg)"
-                />
-                <ModalContent bg="rgba(255, 255, 255, 0.8)" /* semi-transparent background color */
-                    backdropFilter="blur(10px)" /* for Chrome */
-                    _dark={{ bg: 'rgba(0, 0, 0, 0.8)' }} maxW={{ base: '300px', md: '540px' }} >
-                    <SearchBarModal />
-                    <Flex maxH="80vh" flexDir={'column'} gap={3} overflowY={'scroll'}>
-                        {loading ?
-                         <Flex justifyContent={'center'}>
-                            <CircularProgress color={'blackAlpha.800'}/>
-                        </Flex> : 
-                        searchResult?.map((result) => (
-                            <>
-                                <RLink onClick={onClose} to={`/anime/${result.id}`}>
-                                    <Flex gap={3} px={4} py={2}>
-                                        <Image src={result.image} alt={result.title?.romaji} height={100} width={'16'} fit={'cover'} />
-                                        <Flex flexDir={'column'}>
-                                            <Text fontSize={'md'} fontWeight={'semibold'} noOfLines={1}>{result.title?.romaji}</Text>
-                                            <Text fontSize={'sm'} fontWeight={'medium'} noOfLines={1}>{result.releaseDate}</Text>
-                                        </Flex>
-
-                                    </Flex>
-                                    <Divider></Divider>
-                                </RLink>
-                            </>
-
-                        ))}
-                    </Flex>
-
-                </ModalContent>
-            </Modal>
-        );
-    }
-
-    function ConstructionBox() {
-        return (
-            <Box p={{ base: 8, md: 8 }} borderRadius="md">
-                <Flex flexDirection={'column'} gap={6}>
-                    <Box>
-                        <Flex alignItems="center" mb={4}>
-                            <Icon as={FaTools} fontSize="2xl" mr={2} />
-                            <Heading size={{ base: 'md', md: 'md' }}>Seacrh Feature Under Construction</Heading>
-                        </Flex>
-                        <Text fontSize="lg">
-                            We are currently working hard to bring you this search feature. Stay tuned for
-                            updates!
-                        </Text>
-                    </Box>
-                    <Button onClick={onClose} >
-                        Got It
-                    </Button>
-                </Flex>
-
-            </Box>
-        );
-    };
-
-
-
 }
-
 
 export default MyHeader;
